services/embeddings-engine/requirements.txt
# Embeddings Engine Service - Dependencies
# Generates vector embeddings using AWS Bedrock API

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# AWS Bedrock for embeddings
boto3==1.33.0
botocore==1.33.0

# Alternative embedding providers (optional)
openai==1.3.7
cohere==4.37

# Vector operations
numpy==1.26.2
sentence-transformers==2.2.2

# Qdrant vector database client
qdrant-client==1.7.0

# Utilities
pydantic==2.5.2
python-dotenv==1.0.0

httpx==0.25.2

psycopg2-binary==2.9.9